test: restruct asset order lockscript tests
